fix: frame time should be int

+++ utils/counter.py
@@ -2,8 +2,8 @@ class Counter:
     value: int
 
     def __init__(self, slot=1, value=0):
-        self.slot = slot
-        self.value = slot * value
+        self.slot = int(slot)
+        self.value = int(slot * value)
 
     def increase(self, step=1):
         self.value += step
@@ -12,7 +12,7 @@ class Counter:
         self.value -= step
 
     def reset(self, value=0):
-        self.value = self.slot * value
+        self.value = int(self.slot * value)
 
     def is_left(self) -> bool:
         return self.value > 0
